# services/horizon: Exclude trades with >10% rounding slippage from trade aggregations