--- a/exp/orderbook/pools.go
+++ b/exp/orderbook/pools.go
@@ -67,10 +67,10 @@ func makeTrade(
 	var result xdr.Int64
 	switch tradeType {
 	case tradeTypeDeposit:
-		result, ok = calculatePoolPayout(X, Y, amount, details.Params.Fee)
+		result, _, ok = CalculatePoolPayout(X, Y, amount, details.Params.Fee)
 
 	case tradeTypeExpectation:
-		result, ok = calculatePoolExpectation(X, Y, amount, details.Params.Fee)
+		result, _, ok = CalculatePoolExpectation(X, Y, amount, details.Params.Fee)
 
 	default:
 		return 0, errBadTradeType
@@ -84,65 +84,83 @@ func makeTrade(
 	return result, nil
 }
 
-// calculatePoolPayout calculates the amount of `reserveB` disbursed from the
+// CalculatePoolPayout calculates the amount of `reserveB` disbursed from the
 // pool for a `received` amount of `reserveA` . From CAP-38:
 //
 //      y = floor[(1 - F) Yx / (X + x - Fx)]
 //
 // It returns false if the calculation overflows.
-func calculatePoolPayout(reserveA, reserveB, received xdr.Int64, feeBips xdr.Int32) (xdr.Int64, bool) {
+func CalculatePoolPayout(reserveA, reserveB, received xdr.Int64, feeBips xdr.Int32) (xdr.Int64, xdr.Int64, bool) {
 	X, Y := uint256.NewInt(uint64(reserveA)), uint256.NewInt(uint64(reserveB))
 	F, x := uint256.NewInt(uint64(feeBips)), uint256.NewInt(uint64(received))
+	S := new(uint256.Int) // Rounding Slippage in bips
 
 	// would this deposit overflow the reserve?
 	if received > math.MaxInt64-reserveA {
-		return 0, false
+		return 0, 0, false
 	}
 
-	// We do all of the math in bips, so it's all upscaled by this value.
+	// We do all of the math with 4 extra decimal places of precision, so it's
+	// all upscaled by this value.
 	maxBips := uint256.NewInt(10000)
 	f := new(uint256.Int).Sub(maxBips, F) // upscaled 1 - F
 
 	// right half: X + (1 - F)x
 	denom := X.Mul(X, maxBips).Add(X, new(uint256.Int).Mul(x, f))
 	if denom.IsZero() { // avoid div-by-zero panic
-		return 0, false
+		return 0, 0, false
 	}
 
 	// left half, a: (1 - F) Yx
 	numer := Y.Mul(Y, x).Mul(Y, f)
 
 	// divide & check overflow
-	result := numer.Div(numer, denom)
+	result, rem := new(uint256.Int), new(uint256.Int)
+	result.Div(numer, denom)
+	rem.Mod(numer, denom)
+
+	if !rem.IsZero() {
+		// Recalculate with more precision
+		unrounded, rounded := new(uint256.Int), new(uint256.Int)
+		unrounded.Mul(numer, maxBips).Div(unrounded, denom)
+		rounded.Mul(result, maxBips)
+		S.Sub(unrounded, rounded)
+		S.Abs(S).Mul(S, maxBips)
+		S.Div(S, unrounded)
+	}
 
 	val := xdr.Int64(result.Uint64())
-	return val, result.IsUint64() && val >= 0
+	roundingSlippageBips := xdr.Int64(S.Uint64())
+	ok := result.IsUint64() && val >= 0 && S.IsUint64() && roundingSlippageBips >= 0
+	return val, roundingSlippageBips, ok
 }
 
-// calculatePoolExpectation determines how much of `reserveA` you would need to
+// CalculatePoolExpectation determines how much of `reserveA` you would need to
 // put into a pool to get the `disbursed` amount of `reserveB`.
 //
 //      x = ceil[Xy / ((Y - y)(1 - F))]
 //
 // It returns false if the calculation overflows.
-func calculatePoolExpectation(
+func CalculatePoolExpectation(
 	reserveA, reserveB, disbursed xdr.Int64, feeBips xdr.Int32,
-) (xdr.Int64, bool) {
+) (xdr.Int64, xdr.Int64, bool) {
 	X, Y := uint256.NewInt(uint64(reserveA)), uint256.NewInt(uint64(reserveB))
 	F, y := uint256.NewInt(uint64(feeBips)), uint256.NewInt(uint64(disbursed))
+	S := new(uint256.Int) // Rounding Slippage in bips
 
 	// sanity check: disbursing shouldn't underflow the reserve
 	if disbursed >= reserveB {
-		return 0, false
+		return 0, 0, false
 	}
 
-	// We do all of the math in bips, so it's all upscaled by this value.
-	maxBips := uint256.NewInt(10000)
+	// We do all of the math with 4 extra decimal places of precision, so it's
+	// all upscaled by this value.
+	maxBips := uint256.NewInt(10_000)
 	f := new(uint256.Int).Sub(maxBips, F) // upscaled 1 - F
 
 	denom := Y.Sub(Y, y).Mul(Y, f) // right half: (Y - y)(1 - F)
 	if denom.IsZero() {            // avoid div-by-zero panic
-		return 0, false
+		return 0, 0, false
 	}
 
 	numer := X.Mul(X, y).Mul(X, maxBips) // left half: Xy
@@ -154,10 +172,20 @@ func calculatePoolExpectation(
 	// hacky way to ceil(): if there's a remainder, add 1
 	if !rem.IsZero() {
 		result.AddUint64(result, 1)
+
+		// Recalculate with more precision
+		unrounded, rounded := new(uint256.Int), new(uint256.Int)
+		unrounded.Mul(numer, maxBips).Div(unrounded, denom)
+		rounded.Mul(result, maxBips)
+		S.Sub(unrounded, rounded)
+		S.Abs(S).Mul(S, maxBips)
+		S.Div(S, unrounded)
 	}
 
 	val := xdr.Int64(result.Uint64())
-	return val, result.IsUint64() && val >= 0
+	roundingSlippageBips := xdr.Int64(S.Uint64())
+	ok := result.IsUint64() && val >= 0 && S.IsUint64() && roundingSlippageBips >= 0
+	return val, roundingSlippageBips, ok
 }
 
 // getOtherAsset returns the other asset in the liquidity pool. Note that