--- a/ingest/ledger_transaction.go
+++ b/ingest/ledger_transaction.go
@@ -86,7 +86,7 @@ func (t *LedgerTransaction) GetChanges() ([]Change, error) {
 // GetOperation returns an operation by index.
 func (t *LedgerTransaction) GetOperation(index uint32) (xdr.Operation, bool) {
 	ops := t.Envelope.Operations()
-	if len(ops) == 0 || int(index) >= len(ops) {
+	if int(index) >= len(ops) {
 		return xdr.Operation{}, false
 	}
 	return ops[index], true
@@ -125,7 +125,7 @@ func (t *LedgerTransaction) GetOperationChanges(operationIndex uint32) ([]Change
 }
 
 func operationChanges(ops []xdr.OperationMeta, index uint32) []Change {
-	if len(ops) == 0 || int(index) >= len(ops) {
+	if int(index) >= len(ops) {
 		return []Change{}
 	}