--- a/services/horizon/cmd/db.go
+++ b/services/horizon/cmd/db.go
@@ -396,6 +396,7 @@ func runDBReingestRange(ledgerRanges []history.LedgerRange, reingestForce bool, 
 		CaptiveCoreStoragePath:      config.CaptiveCoreStoragePath,
 		StellarCoreCursor:           config.CursorName,
 		StellarCoreURL:              config.StellarCoreURL,
+		RoundingSlippageFilter:      config.RoundingSlippageFilter,
 	}
 
 	if !ingestConfig.EnableCaptiveCore {