--- a/services/horizon/internal/db2/history/mock_q_trades.go
+++ b/services/horizon/internal/db2/history/mock_q_trades.go
@@ -32,8 +32,8 @@ func (m *MockQTrades) NewTradeBatchInsertBuilder(maxBatchSize int) TradeBatchIns
 	return a.Get(0).(TradeBatchInsertBuilder)
 }
 
-func (m *MockQTrades) RebuildTradeAggregationBuckets(ctx context.Context, fromLedger, toLedger uint32) error {
-	a := m.Called(ctx, fromLedger, toLedger)
+func (m *MockQTrades) RebuildTradeAggregationBuckets(ctx context.Context, fromLedger, toLedger uint32, roundingSlippageFilter float64) error {
+	a := m.Called(ctx, fromLedger, toLedger, roundingSlippageFilter)
 	return a.Error(0)
 }