--- a/services/horizon/internal/db2/history/trade_batch_insert_builder.go
+++ b/services/horizon/internal/db2/history/trade_batch_insert_builder.go
@@ -2,6 +2,7 @@ package history
 
 import (
 	"context"
+	"math/big"
 	"time"
 
 	"github.com/guregu/null"
@@ -47,6 +48,8 @@ type InsertTrade struct {
 
 	PriceN int64 `db:"price_n"`
 	PriceD int64 `db:"price_d"`
+
+	RoundingSlippage *big.Rat `db:"rounding_slippage"`
 }
 
 // TradeBatchInsertBuilder is used to insert trades into the