--- a/services/horizon/internal/db2/history/trade_batch_insert_builder.go
+++ b/services/horizon/internal/db2/history/trade_batch_insert_builder.go
@@ -49,6 +49,8 @@ type InsertTrade struct {
 	PriceD int64 `db:"price_d"`
 
 	RoundingSlippage db.NullRat `db:"rounding_slippage"`
+	BaseReserves     null.Int   `db:"base_reserves"`
+	CounterReserves  null.Int   `db:"counter_reserves"`
 }
 
 // TradeBatchInsertBuilder is used to insert trades into the