--- a/services/horizon/internal/db2/schema/migrations/53_add_trades_rounding_slippage.sql
+++ b/services/horizon/internal/db2/schema/migrations/53_add_trades_rounding_slippage.sql
@@ -1,0 +1,7 @@
+-- +migrate Up
+
+-- Add is_dust to trades table
+ALTER TABLE history_trades ADD rounding_slippage numeric(1000, 7);
+
+-- +migrate Down
+ALTER TABLE history_trades DROP rounding_slippage;