@2opremio 2022-01-13
I wonder whether it would be feasible to compute the slippage on the fly instead of storing it in the DB.