@paulbellamy 2022-01-13
The issue with on-the-fly, is that you need the ledger meta, as it contains the pool reserves.