@bartekn 2022-01-21
The code tested in stg (commit 8f924e1781dfc0bccc7e8c7e1287944008db37f9 which isn't part of this PR but letting you know anyway) was crashing on ledger 39270855 with:
```
panic: division by zero
goroutine 121 [running]:
math/big.(*Rat).Quo(0xc0004eaf28, 0xc0004eb0a8, 0xc0004eafe8)
        math/big/rat.go:535 +0x136
github.com/stellar/go/services/horizon/internal/ingest/processors.(*TradeProcessor).roundingSlippage(0xbf2f077a6b1a7e7, {0x7a,>
        github.com/stellar/go/services/horizon/internal/ingest/processors/trades_processor.go:233 +0x7dd
github.com/stellar/go/services/horizon/internal/ingest/processors.(*TradeProcessor).extractTrades(_, {{0xca, 0x5f, 0x24, 0xc8,>
        github.com/stellar/go/services/horizon/internal/ingest/processors/trades_processor.go:375 +0x188d
github.com/stellar/go/services/horizon/internal/ingest/processors.(*TradeProcessor).ProcessTransaction(0xc012db65a0, {0x12a75b>
        github.com/stellar/go/services/horizon/internal/ingest/processors/trades_processor.go:53 +0x158
github.com/stellar/go/services/horizon/internal/ingest.groupTransactionProcessors.ProcessTransaction({{0xc0043c3950, 0x6d6d030>
        github.com/stellar/go/services/horizon/internal/ingest/group_processors.go:67 +0x133
github.com/stellar/go/services/horizon/internal/ingest/processors.StreamLedgerTransactions({0x12a75b0, 0xc000648640}, {0x128e5>
        github.com/stellar/go/services/horizon/internal/ingest/processors/change_processors.go:32 +0x245
github.com/stellar/go/services/horizon/internal/ingest.(*ProcessorRunner).RunTransactionProcessorsOnLedger(_, {_, _})
        github.com/stellar/go/services/horizon/internal/ingest/processor_runner.go:296 +0x328
github.com/stellar/go/services/horizon/internal/ingest.(*ProcessorRunner).RunAllProcessorsOnLedger(_, {_, _})
        github.com/stellar/go/services/horizon/internal/ingest/processor_runner.go:335 +0x406
github.com/stellar/go/services/horizon/internal/ingest.resumeState.run({0xce0d8f}, 0xc000032480)
        github.com/stellar/go/services/horizon/internal/ingest/fsm.go:471 +0x13fe
github.com/stellar/go/services/horizon/internal/ingest.(*system).runStateMachine(0xc000032480, {0x129c730, 0x19adaa0})
        github.com/stellar/go/services/horizon/internal/ingest/main.go:580 +0x364
github.com/stellar/go/services/horizon/internal/ingest.(*system).Run(0x0)
        github.com/stellar/go/services/horizon/internal/ingest/main.go:478 +0x45
github.com/stellar/go/services/horizon/internal.(*App).Serve.func1()
        github.com/stellar/go/services/horizon/internal/app.go:102 +0x35
created by github.com/stellar/go/services/horizon/internal.(*App).Serve
        github.com/stellar/go/services/horizon/internal/app.go:101 +0x291
```