@paulbellamy 2022-01-25
With orderbook dumped to ledger `39339519`...

master as of `4920b27104df0ec7631eae28b5295e220fb59be7`
```
$ go test ./exp/orderbook -bench '.*' -offers offers.dump -pools pools.dump
goos: darwin
goarch: arm64
pkg: github.com/stellar/go/exp/orderbook
BenchmarkVibrantPath-8                               160           7442436 ns/op         2868812 B/op         44 allocs/op
BenchmarkMultipleDestinationAssets-8                 100          10807999 ns/op         2871080 B/op         70 allocs/op
BenchmarkTestData-8                                    2         735662062 ns/op        286842320 B/op      2299 allocs/op
BenchmarkSingleLiquidityPoolExchange/deposit-8          1000000000               0.0000003 ns/op
BenchmarkSingleLiquidityPoolExchange/exchange-8         1000000000               0.0000001 ns/op
BenchmarkLiquidityPoolDeposits-8                        17120752                69.89 ns/op            0 B/op          0 allocs/op
BenchmarkLiquidityPoolExpectations-8                    71820445                16.11 ns/op            0 B/op          0 allocs/op
PASS
ok      github.com/stellar/go/exp/orderbook     32.008s
```

This branch as of `2af05facc58b59e955ae0e5fc59aa89aab190cea`
```
$ go test ./exp/orderbook -bench '.*' -offers offers.dump -pools pools.dump
goos: darwin
goarch: arm64
pkg: github.com/stellar/go/exp/orderbook
BenchmarkVibrantPath-8                               153           7868738 ns/op         2868812 B/op         44 allocs/op
BenchmarkMultipleDestinationAssets-8                 100          10674957 ns/op         2871083 B/op         70 allocs/op
BenchmarkTestData-8                                    2         745304916 ns/op        286842392 B/op      2299 allocs/op
BenchmarkSingleLiquidityPoolExchange/deposit-8          1000000000               0.0000004 ns/op
BenchmarkSingleLiquidityPoolExchange/exchange-8         1000000000               0.0000001 ns/op
BenchmarkLiquidityPoolDeposits-8                         7928089               152.8 ns/op             0 B/op          0 allocs/op
BenchmarkLiquidityPoolExpectations-8                    71751022                16.35 ns/op            0 B/op          0 allocs/op
PASS
ok      github.com/stellar/go/exp/orderbook     32.745s
```

So, the main difference is in BenchmarkLiquidityPoolDeposits. (Could be that the Expectations scenario isn't hitting the rounding slippage), which is ~2x slower. Not ideal. I'm not really sure how to write the math faster than it is, so maybe someone else knows, or the other option would be to only store the reserves, and calculate it as part of rebuilding trade aggregations. Which would be less efficient, but not in the hot path.