@paulbellamy 2022-01-25
Ok, disabled the roundingSlippage when called through /paths, and tidied up some allocations. so benchmark is now the same as master.