@paulbellamy 2022-02-03
Uncovered a bug in effect ingestion while testing this, so this is blocked on https://github.com/stellar/go/issues/4203